+++ database/migrations/2018_02_11_134915_create_loans_table.php
@@ -21,7 +21,7 @@ class CreateLoansTable extends Migration
                 ->on('users')
                 ->onDelete('cascade')
                 ->onupdate('cascade');
-            $table->unsignedInteger('loan')->nullable();
+            $table->BigInteger('loan')->nullable();
             $table->string('description')->nullable();
             $table->dateTime('date_time');
             $table->boolean('force')->default('0');
